+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Introduced save state file version 12.
 - GameDB: Force fast bus timings to fix crashes in Deep Fear. (#740)
 - Input: Fixed analog to D-Pad axis conversion to not overwrite whenever an input was released in opposite direction. (#754; @PringleElUno)
 - MIDI: Defend against crashes when the library fails to initialize.
+- SCU: Timer fixes. (thanks to @celeriyacon)
 - SH2: Fix `@(disp.PC)` loads being decoded as stores for watchpoints.
 - SH2: Fix `ldc/lds @Rm` decoding from the wrong opcode bits for watchpoints.
 - VDP1: Fix handling of zero horizontal character size in CMDSIZE.